+++ src/main/java/de/ait/smallBusiness_be/products/dao/ProductCategoryRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 /**
  * SmallBusiness_BE
@@ -16,7 +15,7 @@ import java.util.Optional;
  */
 
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
 
     List<ProductCategory> findByNameIgnoreCaseOrArtNameIgnoreCase(String name, String artName);
 }
