+++ src/main/java/de/ait/smallBusiness_be/products/service/ProductCategoryService.java
@@ -15,9 +15,9 @@ import java.util.List;
 
 public interface ProductCategoryService {
     ProductCategoryDto addProductCategory(NewProductCategoryDto newProductCategoryDto);
-    ProductCategoryDto getProductCategoryById(Integer id);
-    void deleteProductCategoryById(Integer id);
-    ProductCategoryDto updateProductCategory(Integer id, NewProductCategoryDto newProductCategoryDto);
+    ProductCategoryDto getProductCategoryById(Long id);
+    void deleteProductCategoryById(Long id);
+    ProductCategoryDto updateProductCategory(Long id, NewProductCategoryDto newProductCategoryDto);
     List<ProductCategoryDto> findAllProductCategories();
 
 }
